Tier promotion/demotion, misc. additions/fixes.
SaltyStats will now keep track of character promotions and demotions. A
character's win/loss records in each tier are now independent of each
other. This involved a change in data design, so statfiles from 0.1.0
will not work with 0.1.1.

The stats display will now show a character's win rate percentage for
each tier they have fought in, as well as the number of matches won/lost
in each tier.

SaltyStats will only keep track of fights that take place in matchmaking
mode. Tournaments might get added to the records, but I am not planning
on tracking exhibitions.

Fixed a bug that caused the roster size to be measured incorrectly.