Added fight history.
SaltyStats will now record the winners of every fight. When a fight
stats, SaltyStats will check to see if it is a rematch and it will
display the outcomes of any previous matchups.

Fight records are saved in the same statfile as character records. Your
old statfile is compatible with SaltyStats 0.3; it will be upgraded
automatically when you run 0.3 for the first time. After doing so, it
will no longer be compatible with earlier versions of SaltyStats.

You can now search the fight history using SaltyStats Manager. Also
added a "reload" function to reload the statfile.